feat: refactor of mass destruction - slack events API
For new apps slack recommends that the events API should be used.
This new API is based on webhooks that its URLs are registered on the
App's settings page via slack website and when an event happens,
slack will call the informed URL with the event payload.

This commit also ditches fastify and uses express as the acting
webserver, since the slack API does not looks to play nice with fastify. :/