feat: add "do not ask me" checkbox for not added emojis
When an user sends a reward containing an emoji that is not
allowed for the team, Pinata-Libre will ask the user if he
wants to add it to the allowed list. However, this process
can be annoying depending on the amount of messages
the channel has.
In order to avoid such problems, this commit adds a checkbox
that will make the pinata-libre to stop asking the
user if he wants to add the emoji to the allowed list.